Remove useAuthStateEffect and rename redux_hooks
