+++ src/hooks/index.ts
@@ -1,2 +1,1 @@
-export { useAppDispatch, useAppSelector } from './redux_hooks';
-export { useAuthStateEffect } from './useAuthStateEffect';
+export { useAppDispatch, useAppSelector } from './redux';
